+++ samples/Sentry.Samples.Console.Basic/Program.cs
@@ -1,7 +1,13 @@
 using Sentry;
 
-using (SentrySdk.Init("https://eb18e953812b41c3aeb042e666fd3b5c@o447951.ingest.sentry.io/5428537"))
+using var _ = SentrySdk.Init(o =>
 {
-    // The following exception is captured and sent to Sentry
-    throw null;
-}
+    // The DSN is required.
+    o.Dsn = "https://eb18e953812b41c3aeb042e666fd3b5c@o447951.ingest.sentry.io/5428537";
+
+    // When debug is enabled, the Sentry client will emit detailed debugging information to the console.
+    o.Debug = true;
+});
+
+// The following unhandled exception will be captured and sent to Sentry.
+throw new Exception("test");