+++ src/Sentry/HubExtensions.cs
@@ -155,6 +155,12 @@ public static class HubExtensions
         public void Dispose() => _scope.Dispose();
     }
 
+    internal static SentryId CaptureExceptionInternal(this IHub hub, Exception ex) =>
+        hub.CaptureEventInternal(new SentryEvent(ex));
+
+    internal static SentryId CaptureEventInternal(this IHub hub, SentryEvent evt) =>
+        hub is IHubEx hubEx ? hubEx.CaptureEventInternal(evt) : hub.CaptureEvent(evt);
+
     /// <summary>
     /// Captures the exception with a configurable scope callback.
     /// </summary>
@@ -162,10 +168,8 @@ public static class HubExtensions
     /// <param name="ex">The exception.</param>
     /// <param name="configureScope">The callback to configure the scope.</param>
     /// <returns>The Id of the event</returns>
-    public static SentryId CaptureException(this IHub hub, Exception ex, Action<Scope> configureScope)
-        => !hub.IsEnabled
-            ? new SentryId()
-            : hub.CaptureEvent(new SentryEvent(ex), configureScope);
+    public static SentryId CaptureException(this IHub hub, Exception ex, Action<Scope> configureScope) =>
+        hub.CaptureEvent(new SentryEvent(ex), configureScope);
 
     /// <summary>
     /// Captures a message with a configurable scope callback.
@@ -175,20 +179,22 @@ public static class HubExtensions
     /// <param name="configureScope">The callback to configure the scope.</param>
     /// <param name="level">The message level.</param>
     /// <returns>The Id of the event</returns>
-    public static SentryId CaptureMessage(
-        this IHub hub,
-        string message,
-        Action<Scope> configureScope,
+    public static SentryId CaptureMessage(this IHub hub, string message, Action<Scope> configureScope,
         SentryLevel level = SentryLevel.Info)
-        => !hub.IsEnabled || string.IsNullOrWhiteSpace(message)
-            ? new SentryId()
-            : hub.CaptureEvent(
-                new SentryEvent
-                {
-                    Message = message,
-                    Level = level
-                },
-                configureScope);
+    {
+        if (string.IsNullOrWhiteSpace(message))
+        {
+            return new SentryId();
+        }
+
+        var sentryEvent = new SentryEvent
+        {
+            Message = message,
+            Level = level
+        };
+
+        return hub.CaptureEvent(sentryEvent, configureScope);
+    }
 
     internal static ITransaction StartTransaction(
         this IHub hub,