+++ src/Sentry/Internal/IHubEx.cs
@@ -1,0 +1,6 @@
+namespace Sentry.Internal;
+
+internal interface IHubEx : IHub
+{
+    SentryId CaptureEventInternal(SentryEvent evt, Scope? scope = null);
+}